Move ``TestVariablesChecker`` to functional tests
