+++ tests/checkers/unittest_variables.py
@@ -23,14 +23,12 @@
 import os
 import re
 import sys
-import unittest
 from pathlib import Path
 
 import astroid
 
 from pylint.checkers import variables
-from pylint.constants import IS_PYPY
-from pylint.interfaces import HIGH, UNDEFINED
+from pylint.interfaces import HIGH
 from pylint.testutils import CheckerTestCase, MessageTest, linter, set_config
 
 REGR_DATA_DIR = str(Path(__file__).parent / ".." / "regrtest_data")
@@ -40,35 +38,6 @@ class TestVariablesChecker(CheckerTestCase):
 
     CHECKER_CLASS = variables.VariablesChecker
 
-    def test_bitbucket_issue_78(self) -> None:
-        """Issue 78 report a false positive for unused-module"""
-        module = astroid.parse(
-            """
-        from sys import path
-        path += ['stuff']
-        def func():
-            other = 1
-            return len(other)
-        """
-        )
-        with self.assertNoMessages():
-            self.walk(module)
-
-    @set_config(ignored_modules=("argparse",))
-    def test_no_name_in_module_skipped(self) -> None:
-        """Make sure that 'from ... import ...' does not emit a
-        'no-name-in-module' with a module that is configured
-        to be ignored.
-        """
-
-        node = astroid.extract_node(
-            """
-        from argparse import THIS_does_not_EXIST
-        """
-        )
-        with self.assertNoMessages():
-            self.checker.visit_importfrom(node)
-
     def test_all_elements_without_parent(self) -> None:
         node = astroid.extract_node("__all__ = []")
         node.value.elts.append(astroid.Const("test"))
@@ -76,142 +45,6 @@ class TestVariablesChecker(CheckerTestCase):
         with self.assertNoMessages():
             self.checker.visit_module(root)
             self.checker.leave_module(root)
-
-    def test_redefined_builtin_ignored(self) -> None:
-        node = astroid.parse(
-            """
-        from future.builtins import open
-        """
-        )
-        with self.assertNoMessages():
-            self.checker.visit_module(node)
-
-    @set_config(redefining_builtins_modules=("os",))
-    def test_redefined_builtin_custom_modules(self) -> None:
-        node = astroid.parse(
-            """
-        from os import open
-        """
-        )
-        with self.assertNoMessages():
-            self.checker.visit_module(node)
-
-    @set_config(redefining_builtins_modules=("os",))
-    def test_redefined_builtin_modname_not_ignored(self) -> None:
-        node = astroid.parse(
-            """
-        from future.builtins import open
-        """
-        )
-        with self.assertAddsMessages(
-            MessageTest("redefined-builtin", node=node.body[0], args="open")
-        ):
-            self.checker.visit_module(node)
-
-    @set_config(redefining_builtins_modules=("os",))
-    def test_redefined_builtin_in_function(self) -> None:
-        node = astroid.extract_node(
-            """
-        def test():
-            from os import open
-        """
-        )
-        with self.assertNoMessages():
-            self.checker.visit_module(node.root())
-            self.checker.visit_functiondef(node)
-
-    def test_unassigned_global(self) -> None:
-        node = astroid.extract_node(
-            """
-            def func():
-                global sys  #@
-                import sys, lala
-        """
-        )
-        msg = MessageTest("global-statement", node=node, confidence=UNDEFINED)
-        with self.assertAddsMessages(msg):
-            self.checker.visit_global(node)
-
-    def test_listcomp_in_decorator(self) -> None:
-        """Make sure class attributes in scope for listcomp in decorator.
-
-        https://github.com/PyCQA/pylint/issues/511
-        """
-        module = astroid.parse(
-            """
-        def dec(inp):
-            def inner(func):
-                print(inp)
-                return func
-            return inner
-
-
-        class Cls:
-
-            DATA = "foo"
-
-            @dec([x for x in DATA])
-            def fun(self):
-                pass
-        """
-        )
-        with self.assertNoMessages():
-            self.walk(module)
-
-    def test_listcomp_in_ancestors(self) -> None:
-        """Ensure list comprehensions in base classes are scoped correctly
-
-        https://github.com/PyCQA/pylint/issues/3434
-        """
-        module = astroid.parse(
-            """
-        import collections
-
-
-        l = ["a","b","c"]
-
-
-        class Foo(collections.namedtuple("Foo",[x+"_foo" for x in l])):
-            pass
-        """
-        )
-        with self.assertNoMessages():
-            self.walk(module)
-
-    def test_return_type_annotation(self) -> None:
-        """Make sure class attributes in scope for return type annotations.
-
-        https://github.com/PyCQA/pylint/issues/1976
-        """
-        module = astroid.parse(
-            """
-        class MyObject:
-            class MyType:
-                pass
-            def my_method(self) -> MyType:
-                pass
-        """
-        )
-        with self.assertNoMessages():
-            self.walk(module)
-
-    @unittest.skipIf(IS_PYPY, "PyPy does not parse type comments")
-    def test_attribute_in_type_comment(self):
-        """Ensure attribute lookups in type comments are accounted for.
-
-        https://github.com/PyCQA/pylint/issues/4603
-        """
-        module = astroid.parse(
-            """
-        import foo
-        from foo import Bar, Boo
-        a = ... # type: foo.Bar
-        b = ... # type: foo.Bar[Boo]
-        c = ... # type: Bar.Boo
-        """
-        )
-        with self.assertNoMessages():
-            self.walk(module)
 
 
 class TestVariablesCheckerWithTearDown(CheckerTestCase):
