+++ tests/functional/n/no/no_name_in_module.rc
@@ -1,0 +1,2 @@
+[TYPECHECK]
+ignored-modules=argparse
