+++ tests/functional/r/redefined_builtin.rc
@@ -1,0 +1,2 @@
+[VARIABLES]
+redefining-builtins-modules=os
