+++ tests/functional/r/redefined_builtin.txt
@@ -1,2 +1,3 @@
-redefined-builtin:6:4:6:8:function:Redefining built-in 'type':UNDEFINED
-redefined-builtin:10:0:10:3::Redefining built-in 'map':UNDEFINED
+redefined-builtin:9:4:9:8:function:Redefining built-in 'type':UNDEFINED
+redefined-builtin:14:0:14:3::Redefining built-in 'map':UNDEFINED
+redefined-builtin:19:0:19:22::Redefining built-in 'open':UNDEFINED
