+++ tests/functional/u/undefined/undefined_variable_typing.rc
@@ -1,0 +1,2 @@
+[testoptions]
+except_implementations=PyPy
